move file to upper level
