+++ src/components/datagridField/util.js
@@ -173,24 +173,3 @@ export function renderFieldsAndSubsections(
     );
   });
 }
-
-export const applyFieldResolvers = (columnModel, rowData) => {
-  return columnModel.map((column) => {
-    const colModelCopy = _.cloneDeep(columnModel);
-    const { fieldMetaResolver = [] } = column;
-    let meta = column.meta || {};
-
-    for (let i = 0; i < fieldMetaResolver.length; i += 1) {
-      const fieldResolver = fieldMetaResolver[i];
-
-      if (typeof fieldResolver === "function") {
-        meta = fieldResolver(colModelCopy, rowData, column.dataIndex) || meta;
-      }
-    }
-
-    return {
-      ...column,
-      meta,
-    };
-  });
-}
